Add ordinal sdtype
