+++ rdt/transformers/categorical.py
@@ -41,9 +41,6 @@ class UniformEncoder(BaseTransformer):
     Null values are considered just another category.
 
     Args:
-        order_by (str or None):
-            String defining how to order the data before applying the labels. Options are
-            'alphabetical', 'numerical' and ``None``. Defaults to ``None``.
         missing_value_encoding (str or None):
             How to encode missing values. If ``'new_category'``, missing values are encoded as
             their own category. If ``None``, missing values are not encoded and remain missing.
@@ -57,38 +54,10 @@ class UniformEncoder(BaseTransformer):
     dtype = None
     missing_value_encoding = 'new_category'
 
-    def __init__(self, order_by=None, missing_value_encoding='new_category'):
+    def __init__(self, missing_value_encoding='new_category'):
         super().__init__()
-        if order_by not in [None, 'alphabetical', 'numerical_value']:
-            raise TransformerInputError(
-                "order_by must be one of the following values: None, 'numerical_value' or "
-                "'alphabetical'"
-            )
-
         _validate_missing_value_encoding(missing_value_encoding)
-        self.order_by = order_by
         self.missing_value_encoding = missing_value_encoding
-
-    def _order_categories(self, unique_data):
-        nans = pd.isna(unique_data)
-        if self.order_by == 'alphabetical':
-            # pylint: disable=invalid-unary-operand-type
-            if any(not isinstance(item, str) for item in unique_data[~nans]):
-                raise TransformerInputError(
-                    "The data must be of type string if order_by is 'alphabetical'."
-                )
-        elif self.order_by == 'numerical_value':
-            if any(not np.issubdtype(type(item), np.number) for item in unique_data[~nans]):
-                raise TransformerInputError(
-                    "The data must be numerical if order_by is 'numerical_value'."
-                )
-
-        if self.order_by is not None:
-            unique_data = np.sort(unique_data[~nans])  # pylint: disable=invalid-unary-operand-type
-            if nans.any():
-                unique_data = np.append(unique_data, [None])
-
-        return unique_data
 
     @classmethod
     def _get_message_unseen_categories(cls, unseen_categories):
@@ -146,7 +115,6 @@ class UniformEncoder(BaseTransformer):
             data = fill_nan_with_none(data)
 
         labels = pd.unique(data)
-        labels = self._order_categories(labels)
         freq = data.value_counts(normalize=True, dropna=False)
         nan_value = freq[np.nan] if np.nan in freq.index else None
         freq = freq.reindex(labels, fill_value=nan_value).array
@@ -256,127 +224,6 @@ class UniformEncoder(BaseTransformer):
         result = try_convert_to_dtype(result, self.dtype)
 
         return result
-
-
-class OrderedUniformEncoder(UniformEncoder):
-    """Ordered uniform encoder for categorical data.
-
-    This class works very similarly to the ``UniformEncoder``, except that it requires the ordering
-    for the labels to be provided.
-    Null values are considered just another category.
-
-    Args:
-        order (list):
-            A list of all the unique categories for the data. The order of the list determines the
-            label that each category will get.
-        missing_value_encoding (str or None):
-            How to encode missing values. If ``'new_category'``, missing values are encoded as
-            their own category. If ``None``, missing values are not encoded and remain missing.
-            Defaults to ``'new_category'``.
-    """
-
-    def __init__(self, order, missing_value_encoding='new_category'):
-        _validate_missing_value_encoding(missing_value_encoding)
-        self.missing_value_encoding = missing_value_encoding
-        self.order = fill_nan_with_none(pd.Series(order))
-        if not self.order.is_unique:
-            error_msg = (
-                "The OrderedUniformEncoder has duplicate categories in the 'order' parameter. "
-                'Please drop the duplicates to proceed.'
-            )
-            raise TransformerInputError(error_msg)
-
-        super().__init__(missing_value_encoding=missing_value_encoding)
-
-    def __repr__(self):
-        """Represent initialization of transformer as text.
-
-        Returns:
-            str:
-                The name of the transformer followed by any non-default parameters.
-        """
-        class_name = self.__class__.get_name()
-        custom_args = ['order=<CUSTOM>']
-        if self.missing_value_encoding != 'new_category':
-            custom_args.append(f'missing_value_encoding={repr(self.missing_value_encoding)}')
-
-        args_string = ', '.join(custom_args)
-        return f'{class_name}({args_string})'
-
-    def _get_order(self):
-        if self.missing_value_encoding is None:
-            return self.order[~pd.isna(self.order)]
-
-        return self.order
-
-    def _check_unknown_categories(self, data):
-        order = self._get_order()
-        unknown = ~data.isin(order)
-        if self.missing_value_encoding is None:
-            unknown &= ~pd.isna(data)
-
-        missing = list(data[unknown].unique())
-        if len(missing) > 0:
-            raise TransformerInputError(
-                f"Unknown categories '{missing}'. All possible categories must be defined in the "
-                "'order' parameter."
-            )
-
-    def _fit(self, data):
-        """Fit the transformer to the data.
-
-        Create all the class attributes while respecting the speicified
-        order of the labels.
-
-        Args:
-            data (pandas.Series):
-                Data to fit the transformer to.
-        """
-        self.dtype = data.dtypes
-        order = self._get_order()
-        if self.missing_value_encoding is None:
-            data = data[~pd.isna(data)]
-        else:
-            data = fill_nan_with_none(data)
-
-        self._check_unknown_categories(data)
-
-        category_not_seen = set(order.dropna()) != set(data.dropna())
-        nans_not_seen = (
-            self.missing_value_encoding == 'new_category'
-            and pd.isna(order).any()
-            and not pd.isna(data).any()
-        )
-        if category_not_seen or nans_not_seen:
-            unseen_categories = [x for x in order if x not in data.array]
-            categories_to_print = self._get_message_unseen_categories(unseen_categories)
-            LOGGER.info(
-                "For column '%s', some of the provided category values were not present in the"
-                ' data during fit: (%s).',
-                self.get_input_column(),
-                categories_to_print,
-            )
-
-            freq = data.value_counts(normalize=True, dropna=False)
-            freq = 0.9 * freq
-            for category in unseen_categories:
-                freq.loc[category] = 0.1 / len(unseen_categories)
-
-        else:
-            freq = data.value_counts(normalize=True, dropna=False)
-
-        nan_value = freq[np.nan] if np.nan in freq.index else None
-        freq = freq.reindex(order, fill_value=nan_value).array
-
-        self.frequencies, self.intervals = self._compute_frequencies_intervals(order, freq)
-
-    def _transform(self, data):
-        """Map the category to a continuous value."""
-        if self.missing_value_encoding == 'new_category':
-            data = fill_nan_with_none(data)
-
-        self._check_unknown_categories(data)
-        return super()._transform(data)
 
 
 class FrequencyEncoder(BaseTransformer):
@@ -790,12 +637,6 @@ class LabelEncoder(BaseTransformer):
         add_noise (bool):
             Whether to generate uniform noise around the label for each category.
             Defaults to ``False``.
-        order_by (None or str):
-            A string defining how to order the categories before assigning them labels. Defaults to
-            ``None``. Options include:
-            - ``'numerical_value'``: Order the categories by numerical value.
-            - ``'alphabetical'``: Order the categories alphabetically.
-            - ``None``: Use the order that the categories appear in when fitting.
         missing_value_encoding (str or None):
             How to encode missing values. If ``'new_category'``, missing values are encoded as
             their own category. If ``None``, missing values are not encoded and remain missing.
@@ -809,42 +650,11 @@ class LabelEncoder(BaseTransformer):
     dtype = 'O'
     missing_value_encoding = 'new_category'
 
-    def __init__(self, add_noise=False, order_by=None, missing_value_encoding='new_category'):
+    def __init__(self, add_noise=False, missing_value_encoding='new_category'):
         super().__init__()
         self.add_noise = add_noise
-        if order_by not in [None, 'alphabetical', 'numerical_value']:
-            raise TransformerInputError(
-                "order_by must be one of the following values: None, 'numerical_value' or "
-                "'alphabetical'"
-            )
-
         _validate_missing_value_encoding(missing_value_encoding)
-        self.order_by = order_by
         self.missing_value_encoding = missing_value_encoding
-
-    def _order_categories(self, unique_data):
-        if len(unique_data) == 0:
-            return unique_data
-
-        if self.order_by == 'alphabetical':
-            if unique_data.dtype.type not in [np.str_, np.object_]:
-                raise TransformerInputError(
-                    "The data must be of type string if order_by is 'alphabetical'."
-                )
-
-        elif self.order_by == 'numerical_value':
-            if not np.issubdtype(unique_data.dtype.type, np.number):
-                raise TransformerInputError(
-                    "The data must be numerical if order_by is 'numerical_value'."
-                )
-
-        if self.order_by is not None:
-            nans = pd.isna(unique_data)
-            unique_data = np.sort(unique_data[~nans])  # pylint: disable=E1130
-            if nans.any():
-                unique_data = np.append(unique_data, [np.nan])
-
-        return unique_data
 
     def _fit(self, data):
         """Fit the transformer to the data.
@@ -864,7 +674,6 @@ class LabelEncoder(BaseTransformer):
             data = data.infer_objects().fillna(np.nan)
 
         unique_data = pd.unique(data)
-        unique_data = self._order_categories(unique_data)
         self.values_to_categories = dict(enumerate(unique_data))
         self.categories_to_values = {
             category: value for value, category in self.values_to_categories.items()
@@ -944,114 +753,3 @@ class LabelEncoder(BaseTransformer):
         data = try_convert_to_dtype(data, self.dtype)
 
         return data
-
-
-class OrderedLabelEncoder(LabelEncoder):
-    """Custom label encoder for categorical data.
-
-    This class works very similarly to the ``LabelEncoder``, except that it requires the ordering
-    for the labels to be provided.
-
-    Null values are considered just another category.
-
-    Args:
-        order (list):
-            A list of all the unique categories for the data. The order of the list determines the
-            label that each category will get.
-        add_noise (bool):
-            Whether to generate uniform noise around the label for each category.
-            Defaults to ``False``.
-        missing_value_encoding (str or None):
-            How to encode missing values. If ``'new_category'``, missing values are encoded as
-            their own category. If ``None``, missing values are not encoded and remain missing.
-            Defaults to ``'new_category'``.
-    """
-
-    def __init__(self, order, add_noise=False, missing_value_encoding='new_category'):
-        _validate_missing_value_encoding(missing_value_encoding)
-        self.missing_value_encoding = missing_value_encoding
-        self.order = pd.Series(order).fillna(np.nan)
-        if not self.order.is_unique:
-            err_msg = (
-                "The OrderedLabelEncoder has duplicate categories in the 'order' parameter. "
-                'Please drop the duplicates to proceed.'
-            )
-            raise TransformerInputError(err_msg)
-
-        super().__init__(add_noise=add_noise, missing_value_encoding=missing_value_encoding)
-
-    def __repr__(self):
-        """Represent initialization of transformer as text.
-
-        Returns:
-            str:
-                The name of the transformer followed by any non-default parameters.
-        """
-        class_name = self.__class__.get_name()
-        custom_args = []
-        custom_args.append('order=<CUSTOM>')
-        if self.add_noise:
-            custom_args.append(f'add_noise={self.add_noise}')
-        if self.missing_value_encoding != 'new_category':
-            custom_args.append(f'missing_value_encoding={repr(self.missing_value_encoding)}')
-
-        args_string = ', '.join(custom_args)
-        return f'{class_name}({args_string})'
-
-    def _get_order(self):
-        if self.missing_value_encoding is None:
-            return self.order[~pd.isna(self.order)]
-
-        return self.order
-
-    def _fit(self, data):
-        """Fit the transformer to the data.
-
-        Generate a unique integer representation for each category and
-        store them in the ``categories_to_values`` dict and its reverse
-        ``values_to_categories``.
-
-        Args:
-            data (pandas.Series):
-                Data to fit the transformer to.
-        """
-        self.dtype = data.dtype
-        data = data.infer_objects()
-        if self.missing_value_encoding == 'new_category':
-            data = data.fillna(np.nan)
-
-        order = self._get_order()
-
-        unknown = ~data.isin(order)
-        if self.missing_value_encoding is None:
-            unknown &= ~pd.isna(data)
-
-        missing = list(data[unknown].unique())
-        if len(missing) > 0:
-            raise TransformerInputError(
-                f"Unknown categories '{missing}'. All possible categories must be defined in the "
-                "'order' parameter."
-            )
-
-        if self.missing_value_encoding is None:
-            data = data[~pd.isna(data)]
-
-        self.values_to_categories = dict(enumerate(order))
-        self.categories_to_values = {
-            category: value for value, category in self.values_to_categories.items()
-        }
-
-
-class CustomLabelEncoder(OrderedLabelEncoder):
-    """Deprecated class name for ``OrderedLabelEncoder``.
-
-    Class to ensure backwards compatibility with previous versions of RDT.
-    """
-
-    def __init__(self, order, add_noise=False):
-        warnings.warn(
-            "The 'CustomLabelEncoder' is renamed to 'OrderedLabelEncoder'. Please update the"
-            'name to ensure compatibility with future versions of RDT.',
-            FutureWarning,
-        )
-        super().__init__(order, add_noise)
