+++ tests/integration/transformers/test_categorical.py
@@ -13,8 +13,6 @@ from rdt.transformers import (
     FrequencyEncoder,
     LabelEncoder,
     OneHotEncoder,
-    OrderedLabelEncoder,
-    OrderedUniformEncoder,
     UniformEncoder,
 )
 
@@ -45,7 +43,7 @@ class TestUniformEncoder:
         # Setup
         data = pd.DataFrame({'column_name': [1, 2, 3, 2, 1, 1, 1]})
         column = 'column_name'
-        transformer = UniformEncoder(order_by='numerical_value')
+        transformer = UniformEncoder()
 
         # Run
         transformer.fit(data, column)
@@ -219,57 +217,6 @@ class TestUniformEncoder:
         pd.testing.assert_frame_equal(transform1, transform2)
         with pytest.raises(AssertionError):
             pd.testing.assert_frame_equal(transform1, transform3)
-
-
-class TestOrderedUniformEncoder:
-    """Test class for the OrderedUniformEncoder."""
-
-    def test_order(self):
-        """Test that the ``order`` parameter is respected."""
-        # Setup
-        data = pd.DataFrame({'column_name': [1, 2, 3, 2, np.nan, 1, 1]})
-        transformer = OrderedUniformEncoder(order=[2, 3, np.nan, 1])
-        column = 'column_name'
-
-        # Run
-        transformer.fit(data, column)
-        transformed = transformer.transform(data)
-        reverse = transformer.reverse_transform(transformed)
-        expected_order = pd.Series([2, 3, np.nan, 1], dtype=object)
-
-        # Asserts
-        pd.testing.assert_series_equal(reverse[column], data[column])
-        pd.testing.assert_series_equal(transformer.order, expected_order)
-
-    def test_string(self):
-        """Test that the transformer works with string labels."""
-        # Setup
-        data = pd.DataFrame({'column_name': ['b', 'a', 'c', 'a', np.nan, 'b', 'b']})
-        transformer = OrderedUniformEncoder(order=['a', 'c', np.nan, 'b'])
-        column = 'column_name'
-
-        # Run
-        transformer.fit(data, column)
-        transformed = transformer.transform(data)
-        reverse = transformer.reverse_transform(transformed)
-
-        # Asserts
-        pd.testing.assert_series_equal(reverse[column], data[column])
-
-    def test_mixt_dtype(self):
-        """Test that the transformer works with mixture of dtypes labels."""
-        # Setup
-        data = pd.DataFrame({'column_name': [1, 'a', 'c', 'a', np.nan, 1, 1]})
-        transformer = OrderedUniformEncoder(order=['a', 'c', np.nan, 1])
-        column = 'column_name'
-
-        # Run
-        transformer.fit(data, column)
-        transformed = transformer.transform(data)
-        reverse = transformer.reverse_transform(transformed)
-
-        # Asserts
-        pd.testing.assert_series_equal(reverse[column], data[column])
 
 
 def test_frequency_encoder_numerical_nans():
@@ -648,33 +595,11 @@ def test_label_encoder_numerical_nans_no_warning():
     pd.testing.assert_frame_equal(reverse, data)
 
 
-def test_label_encoder_order_by_numerical():
-    """Test the LabelEncoder appropriately transforms data if `order_by` is 'numerical_value'.
-
-    Input:
-        - pandas.DataFrame of numeric data.
-
-    Output:
-        - Transformed data should map labels to values based on numerical order.
-    """
-
-    data = pd.DataFrame([5, np.nan, 3.11, 100, 67.8, -2.5], columns=['column_name'])
-
-    transformer = LabelEncoder(order_by='numerical_value')
-    transformer.fit(data, 'column_name')
-    transformed = transformer.transform(data)
-    reverse = transformer.reverse_transform(transformed)
-
-    expected = pd.DataFrame([2, 5, 1, 4, 3, 0], columns=['column_name'])
-    pd.testing.assert_frame_equal(transformed, expected)
-    pd.testing.assert_frame_equal(reverse, data)
-
-
 def test_label_encoder_with_numerical_value():
     """Test setting numerical_value works with missing_value_encoding."""
     # Setup
     data = pd.DataFrame({'column_name': pd.Series([2, 1, pd.NA, 3], dtype='object')})
-    transformer = LabelEncoder(order_by='numerical_value', missing_value_encoding=None)
+    transformer = LabelEncoder(missing_value_encoding=None)
 
     # Run
     transformer.fit(data, 'column_name')
@@ -682,33 +607,11 @@ def test_label_encoder_with_numerical_value():
     reverse = transformer.reverse_transform(transformed)
 
     # Assert
-    expected = pd.DataFrame({'column_name': [1.0, 0.0, np.nan, 2.0]})
+    expected = pd.DataFrame({'column_name': [0.0, 1.0, np.nan, 2.0]})
     pd.testing.assert_frame_equal(transformed, expected)
     pd.testing.assert_frame_equal(
         reverse, pd.DataFrame({'column_name': [2, 1, np.nan, 3]}, dtype='object')
     )
-
-
-def test_label_encoder_order_by_alphabetical():
-    """Test the LabelEncoder appropriately transforms data if `order_by` is 'alphabetical'.
-
-    Input:
-        - pandas.DataFrame of string data.
-
-    Output:
-        - Transformed data should map labels to values based on alphabetical order.
-    """
-
-    data = pd.DataFrame(['one', 'two', np.nan, 'three', 'four'], columns=['column_name'])
-
-    transformer = LabelEncoder(order_by='alphabetical')
-    transformer.fit(data, 'column_name')
-    transformed = transformer.transform(data)
-    reverse = transformer.reverse_transform(transformed)
-
-    expected = pd.DataFrame([1, 3, 4, 2, 0], columns=['column_name'])
-    pd.testing.assert_frame_equal(transformed, expected)
-    pd.testing.assert_frame_equal(reverse, data)
 
 
 def test_label_encoder_add_noise():
@@ -753,89 +656,13 @@ def test_label_encoder_with_add_noise_and_missing_value_encoding():
     pd.testing.assert_frame_equal(reverse, pd.DataFrame({'column_name': ['a', np.nan, 'b']}))
 
 
-def test_ordered_label_encoder():
-    """Test the OrderedLabelEncoder end to end.
-
-    Input:
-        - pandas.DataFrame of different types of data.
-
-    Output:
-        - Transformed data should have values based on the provided order.
-        - Reverse transformed data should match the input
-    """
-
-    data = pd.DataFrame(['two', 3, 1, np.nan, 'zero'], columns=['column_name'])
-    transformer = OrderedLabelEncoder(order=['zero', 1, 'two', 3, np.nan])
-    transformer.fit(data, 'column_name')
-
-    transformed = transformer.transform(data)
-    reverse = transformer.reverse_transform(transformed)
-
-    expected = pd.DataFrame([2, 3, 1, 4, 0], columns=['column_name'])
-    pd.testing.assert_frame_equal(transformed, expected)
-    pd.testing.assert_frame_equal(reverse, data)
-
-
-def test_ordered_label_encoder_nans():
-    """The the OrderedLabelEncoder with missing values.
-
-    Input:
-        - pandas.DataFrame of different types of data and different types of missing values.
-
-    Output:
-        - Transformed data should have values based on the provided order.
-        - Reverse transformed data should match the input
-    """
-
-    data = pd.DataFrame(['two', 3, 1, np.nan, 'zero', None], columns=['column_name'])
-    transformer = OrderedLabelEncoder(order=['zero', 1, 'two', 3, None])
-    transformer.fit(data, 'column_name')
-
-    transformed = transformer.transform(data)
-    reverse = transformer.reverse_transform(transformed)
-
-    expected = pd.DataFrame([2, 3, 1, 4, 0, 4], columns=['column_name'])
-    pd.testing.assert_frame_equal(transformed, expected)
-    pd.testing.assert_frame_equal(reverse, data)
-
-
-def test_ordered_label_encoder_numerical_nans_no_warning():
-    """Ensure OrderedLabelEncoder does not emit FutureWarning with nan values.
-
-    Related to Issue #793 (https://github.com/sdv-dev/RDT/issues/793)
-    """
-    # Setup
-    data = pd.DataFrame({'column_name': pd.Series([1, 2, float('nan'), np.nan], dtype='object')})
-    column = 'column_name'
-
-    # Run and Assert
-    transformer = OrderedLabelEncoder(order=[1, 2, np.nan])
-    with warnings.catch_warnings():
-        warnings.simplefilter('error', FutureWarning)
-        transformer.fit(data, column)
-        transformed = transformer.transform(data)
-        reverse = transformer.reverse_transform(transformed)
-
-    pd.testing.assert_frame_equal(reverse, data)
-
-
 categorical_transformers = [
-    UniformEncoder(),
-    OrderedUniformEncoder(order=[1, 'two', 3, 'four', None]),
-    LabelEncoder(),
-    OrderedLabelEncoder(order=[1, 'two', 3, 'four', None]),
+    UniformEncoder(missing_value_encoding=None),
+    LabelEncoder(missing_value_encoding=None),
 ]
 
 
-@pytest.mark.parametrize(
-    'transformer',
-    [
-        UniformEncoder(missing_value_encoding=None),
-        OrderedUniformEncoder(order=[1, 'two', 3, 'four'], missing_value_encoding=None),
-        LabelEncoder(missing_value_encoding=None),
-        OrderedLabelEncoder(order=[1, 'two', 3, 'four'], missing_value_encoding=None),
-    ],
-)
+@pytest.mark.parametrize('transformer', categorical_transformers)
 def test_categorical_transformers_missing_value_encoding_none(transformer):
     """Test categorical transformers preserve missing values when configured to do so."""
     # Setup
@@ -857,15 +684,7 @@ def test_categorical_transformers_missing_value_encoding_none(transformer):
     )
 
 
-@pytest.mark.parametrize(
-    'transformer',
-    [
-        UniformEncoder(missing_value_encoding=None),
-        OrderedUniformEncoder(order=[1, 'two', 3, 'four'], missing_value_encoding=None),
-        LabelEncoder(missing_value_encoding=None),
-        OrderedLabelEncoder(order=[1, 'two', 3, 'four'], missing_value_encoding=None),
-    ],
-)
+@pytest.mark.parametrize('transformer', categorical_transformers)
 def test_categorical_transformers_missing_value_encoding_none_when_nulls_not_seen(
     transformer,
 ):
@@ -893,15 +712,7 @@ def test_categorical_transformers_missing_value_encoding_none_when_nulls_not_see
     )
 
 
-@pytest.mark.parametrize(
-    'transformer',
-    [
-        UniformEncoder(missing_value_encoding=None),
-        OrderedUniformEncoder(order=[1, 'two', 3, 'four'], missing_value_encoding=None),
-        LabelEncoder(missing_value_encoding=None),
-        OrderedLabelEncoder(order=[1, 'two', 3, 'four'], missing_value_encoding=None),
-    ],
-)
+@pytest.mark.parametrize('transformer', categorical_transformers)
 def test_categorical_transformers_missing_value_encoding_none_all_missing(transformer):
     """Test transformers with no learned categories keep missing values missing."""
     # Setup
@@ -918,15 +729,7 @@ def test_categorical_transformers_missing_value_encoding_none_all_missing(transf
     assert reverse_transformed.shape == data.shape
 
 
-@pytest.mark.parametrize(
-    'transformer',
-    [
-        UniformEncoder(missing_value_encoding=None),
-        OrderedUniformEncoder(order=[1, 'two', 3, 'four'], missing_value_encoding=None),
-        LabelEncoder(missing_value_encoding=None),
-        OrderedLabelEncoder(order=[1, 'two', 3, 'four'], missing_value_encoding=None),
-    ],
-)
+@pytest.mark.parametrize('transformer', categorical_transformers)
 def test_categorical_transformers_missing_value_encoding_none_reverse_passes_nulls_through(
     transformer,
 ):
@@ -950,17 +753,11 @@ def test_categorical_transformers_missing_value_encoding_none_reverse_passes_nul
     )
 
 
-@pytest.mark.parametrize(
-    'transformer',
-    [
-        LabelEncoder(missing_value_encoding=None),
-        OrderedLabelEncoder(order=['a', 'b'], missing_value_encoding=None),
-    ],
-)
-def test_label_encoders_missing_value_encoding_none_with_category_dtype(transformer):
+def test_label_encoders_missing_value_encoding_none_with_category_dtype():
     """Test label encoders keep transformed pandas category columns numeric."""
     # Setup
     data = pd.DataFrame({'col': pd.Series(['a', None, 'b'], dtype='category')})
+    transformer = LabelEncoder(missing_value_encoding=None)
 
     # Run
     transformer.fit(data, 'col')
@@ -979,15 +776,10 @@ def test_label_encoders_missing_value_encoding_none_with_category_dtype(transfor
     )
 
 
-@pytest.mark.parametrize(
-    'transformer',
-    [
-        UniformEncoder(),
-        OrderedUniformEncoder(order=[1, 'two', 3, 'four', None]),
-        LabelEncoder(),
-        OrderedLabelEncoder(order=[1, 'two', 3, 'four', None]),
-    ],
-)
+@pytest.mark.parametrize('transformer', [
+    UniformEncoder(),
+    LabelEncoder(),
+])
 def test_categorical_transformers_default_missing_value_encoding_new_category(transformer):
     """Test default missing value handling continues to encode missing as a category."""
     # Setup
