+++ tests/performance/test_performance.py
@@ -8,17 +8,15 @@ from rdt.performance.datasets import get_dataset_generators_by_type
 from rdt.performance.performance import evaluate_transformer_performance
 from rdt.performance.profiling import profile_transformer
 from rdt.transformers import get_transformers_by_type
-from rdt.transformers.categorical import (
-    CustomLabelEncoder,
+from rdt.transformers.numerical import ClusterBasedNormalizer
+from rdt.transformers.ordinal import (
     OrderedLabelEncoder,
     OrderedUniformEncoder,
 )
-from rdt.transformers.numerical import ClusterBasedNormalizer
 
 SANDBOX_TRANSFORMERS = [
     ClusterBasedNormalizer,
     OrderedLabelEncoder,
-    CustomLabelEncoder,
     OrderedUniformEncoder,
 ]
